+++ backend/src/user/dto/add-role.dto.ts
@@ -1,8 +1,6 @@
-import { IsString } from "class-validator";
 import { ApiProperty } from "@nestjs/swagger";
 
 export class AddRoleDto {
-  @ApiProperty({ example: "ADMIN", description: "User's permission" })
-  @IsString({ message: "Must be a string" })
+  @ApiProperty({ example: "admin", description: "User's permission" })
   readonly value: string;
 }