+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+# v3.6.4 (4.05.2022)
+
+## Features:
+
+Frontend:
+1. Added role list for users
+
+# v3.6.3 (4.05.2022)
+
+## Features:
+
+Frontend:
+1. Updated add role functional
+2. Added ban description to users page
+
+# v3.6.2 (3.05.2022)
+
+## Fixes:
+
+Frontend:
+1. Fixed ban/unban user functional
+
+# v3.6.1 (3.05.2022)
+
+## Features:
+
+Frontend:
+1. Updated Readme
+2. Changed folder structures in message
+
 # v3.6.0 (2.05.2022)
 
 ## Features: